Allow using Gradle profiler in cross version tests
