+++ subprojects/build-scan-performance/src/testFixtures/groovy/org/gradle/performance/fixture/BuildScanPerformanceTestRunner.groovy
@@ -27,7 +27,7 @@ import org.gradle.util.GradleVersion
 class BuildScanPerformanceTestRunner extends CrossBuildPerformanceTestRunner {
     private final String pluginCommitSha
 
-    BuildScanPerformanceTestRunner(BuildExperimentRunner experimentRunner, DataReporter<CrossBuildPerformanceResults> dataReporter, String pluginCommitSha, IntegrationTestBuildContext buildContext) {
+    BuildScanPerformanceTestRunner(PerformanceTestBuildExperimentRunner experimentRunner, DataReporter<CrossBuildPerformanceResults> dataReporter, String pluginCommitSha, IntegrationTestBuildContext buildContext) {
         super(experimentRunner, dataReporter, buildContext)
         this.pluginCommitSha = pluginCommitSha
         this.testGroup = "build scan plugin"
