+++ subprojects/performance/performance.gradle
@@ -30,9 +30,16 @@ dependencies {
     performanceTestImplementation libraries.commons_compress.coordinates
     performanceTestImplementation testLibraries.jetty
     performanceTestImplementation(testFixtures(project(":toolingApi")))
+    performanceTestImplementation("org.gradle.profiler:gradle-profiler:1.0-SNAPSHOT") {
+        exclude(group: "com.google.guava", module: "guava")
+    }
 
     // so that all Gradle features are available
     performanceTestRuntimeOnly allTestRuntimeDependencies
+}
+
+configurations.partialDistribution {
+    exclude(group: "org.gradle.profiler", module: "gradle-profiler")
 }
 
 gradlebuildJava {
